Index File With Users Routes
This is the first step of building the app. Starting with the backend, I've wrote a simple express application that uses MongoDB (With no ORM) as database. The server starts by trying to connect to a MongoDB Client, Than imports the users routes that handles all the registration, login, and users retrieval.